requirements.txt
allure_python_commons==2.13.2
Faker==21.0.0
pytest==7.4.3
Requests==2.31.0
